Move data directory creation to thunorctl.py
Data directories for plate-files and downloads were previously
created in settings.py from within the container, but this can
lead to incorrect permissions on the host. Add directory
creation to thunorctl.py on the host and explicitly chmod
to ensure directories are writable. Keep in settings.py
for dev mode/startup check.